Update README.md
The previous link (https://oarc.rutgers.edu/resources/amarel/) is not the user guide but a general info page about Amarel. The correct link to the user guide is https://sites.google.com/view/cluster-user-guide. Even better, we can direct the reader to the specific subsection of the guide ("Connecting to the cluster").