reduce to FS isolation
network isolation for singularity service seems to be a dead end, I did
not manage to get this to work.  The lack of proper network
virtualization in singularity makes this neigh impossible IMHO.  Also,
RMQ is too hard to configure to run in userspace to be worth the effort,
specifically since the respective configuration collide with the MixIn
FS-virtualization approach singularity follows.