don't use `singularity instance`
Singularity instance state is not cleaned up when slurm job completes,
subsequent runs on the same node will conflict with state remnants from
earlier jobs.

This also implies that the started daemons cannot reliably be killed:
singularity does not keep track of spawned processes and there is no
specified way to terminate processes.  We have to rely on slurm to do
the deed...