Commented out throwing away low frequencies in extremesealevel_pointsoverthreshold_project.py
This step was inherited from the SROCC code to prevent 'numerical instabilities', but it introduces artefacts in the return curves at the lowest frequencies, where for some GPD samples these evaluate to z=NaN and for others z=finite, resulting in quantile values of z(f) to be based on only part of the GPD samples. I think this improves if f<1e-6 is not set to NaN except if not supported by the GPD, which is already done. Should not affect AFs as long as f_ref=HCE is used.