Fix for indexing bug in compare_2lm_tmix_cmip6gsat.py
Jan-Erik spotted an issue in the looping over ssps here, which led to using SSP119 instead of SSP126 for emulating GSAT. This is the fix. Which models are dropped by the comparison to simulated GSAT prior to deriving expansion coefficients was not affected so this change has no consequenes further downstream.